# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bin Al-Zain Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin session
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
...
  /admin/customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all registered customers (admin only)
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerProfile"
        "401":
          description: Not authenticated
...
  /auth/register:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "409":
          description: Email already in use
  /auth/login:
    post:
      operationId: loginCustomer
      tags: [auth]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLogin"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current customer session
      responses:
        "200":
          description: Authenticated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "401":
          description: Not authenticated
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update customer profile (name, phone, password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "400":
          description: Validation error
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, nameAr, nameEn, slug, icon]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        icon:
          type: string
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer
        rowNumber:
          type: integer
    CategoryInput:
      type: object
      required: [nameAr, nameEn, slug, icon]
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        icon:
          type: string
        imageUrl:
          type: ["string", "null"]
        rowNumber:
          type: integer
...
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
...
    CustomerProfile:
      type: object
      required: [id, fullName, phone, email, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        createdAt:
          type: string
    CustomerRegistration:
      type: object
      required: [fullName, phone, email, password]
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
    CustomerLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UpdateProfileInput:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        currentPassword:
          type: ["string", "null"]
        newPassword:
          type: ["string", "null"]
...